settings: extract schema defaults and Xfconf binding helpers
