+++ panel-plugin/settings-bindings.h
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2026 Matteo Bonanomi
+ *
+ * This library is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#ifndef WHISKERMENU_SETTINGS_BINDINGS_H
+#define WHISKERMENU_SETTINGS_BINDINGS_H
+
+/* settings-bindings: private translation unit hosting the Xfconf binding
+ * helper classes (Boolean, Integer, String, StringList, SearchActionList).
+ * Their declarations remain in settings.h next to the Settings class they
+ * are friend-coupled to; only the implementations live here.
+ */
+
+#endif // WHISKERMENU_SETTINGS_BINDINGS_H
