+++ panel-plugin/settings.cpp
@@ -19,7 +19,6 @@
 
 #include "command.h"
 #include "plugin.h"
-#include "preset.h"
 #include "search-action.h"
 #include "slot.h"
 
@@ -503,690 +502,6 @@ void Settings::property_changed(const gchar* property, const GValue* value)
 
 //-----------------------------------------------------------------------------
 
-Boolean::Boolean(WhiskerMenu::Settings* settings, const gchar* property, bool data) :
-	m_settings(settings),
-	m_property(property),
-	m_default(data),
-	m_data(m_default)
-{
-}
-
-//-----------------------------------------------------------------------------
-
-void Boolean::load(XfceRc* rc, bool is_default)
-{
-	set(xfce_rc_read_bool_entry(rc, m_property + 1, m_data), !is_default);
-
-	if (is_default)
-	{
-		m_default = m_data;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-bool Boolean::load(const gchar* property, const GValue* value)
-{
-	if (g_strcmp0(m_property, property) != 0)
-	{
-		return false;
-	}
-
-	set(G_VALUE_HOLDS_BOOLEAN(value) ? g_value_get_boolean(value) : m_default, false);
-
-	return true;
-}
-
-//-----------------------------------------------------------------------------
-
-void Boolean::set(bool data, bool store)
-{
-	if (m_data == data)
-	{
-		return;
-	}
-
-	m_data = data;
-
-	if (store && m_settings->channel)
-	{
-		m_settings->begin_property_update();
-		xfconf_channel_set_bool(m_settings->channel, m_property, m_data);
-		m_settings->end_property_update();
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-Integer::Integer(WhiskerMenu::Settings* settings, const gchar* property, int data, int min, int max) :
-	m_settings(settings),
-	m_property(property),
-	m_min(min),
-	m_max(max),
-	m_default(CLAMP(data, min, max)),
-	m_data(m_default)
-{
-}
-
-//-----------------------------------------------------------------------------
-
-void Integer::load(XfceRc* rc, bool is_default)
-{
-	set(xfce_rc_read_int_entry(rc, m_property + 1, m_data), !is_default);
-
-	if (is_default)
-	{
-		m_default = m_data;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-bool Integer::load(const gchar* property, const GValue* value)
-{
-	if (g_strcmp0(m_property, property) != 0)
-	{
-		return false;
-	}
-
-	set(G_VALUE_HOLDS_INT(value) ? g_value_get_int(value) : m_default, false);
-
-	return true;
-}
-
-//-----------------------------------------------------------------------------
-
-void Integer::set(int data, bool store)
-{
-	data = CLAMP(data, m_min, m_max);
-	if (m_data == data)
-	{
-		return;
-	}
-
-	m_data = data;
-
-	if (store && m_settings->channel)
-	{
-		m_settings->begin_property_update();
-		xfconf_channel_set_int(m_settings->channel, m_property, m_data);
-		m_settings->end_property_update();
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-String::String(WhiskerMenu::Settings* settings, const gchar* property, const std::string& data) :
-	m_settings(settings),
-	m_property(property),
-	m_default(data),
-	m_data(m_default)
-{
-}
-
-//-----------------------------------------------------------------------------
-
-void String::load(XfceRc* rc, bool is_default)
-{
-	set(xfce_rc_read_entry(rc, m_property + 1, m_data.c_str()), !is_default);
-
-	if (is_default)
-	{
-		m_default = m_data;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-bool String::load(const gchar* property, const GValue* value)
-{
-	if (g_strcmp0(m_property, property) != 0)
-	{
-		return false;
-	}
-
-	set(G_VALUE_HOLDS_STRING(value) ? g_value_get_string(value) : m_default, false);
-
-	return true;
-}
-
-//-----------------------------------------------------------------------------
-
-void String::set(const std::string& data, bool store)
-{
-	if (m_data == data)
-	{
-		return;
-	}
-
-	m_data = data;
-
-	if (store && m_settings->channel)
-	{
-		m_settings->begin_property_update();
-		xfconf_channel_set_string(m_settings->channel, m_property, m_data.c_str());
-		m_settings->end_property_update();
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-StringList::StringList(WhiskerMenu::Settings* settings, const gchar* property, std::initializer_list<std::string> data) :
-	m_settings(settings),
-	m_property(property),
-	m_default(data),
-	m_data(m_default),
-	m_modified(false),
-	m_saved(false),
-	m_order_unchanged(false)
-{
-}
-
-//-----------------------------------------------------------------------------
-
-int StringList::find(const std::string& value) const
-{
-	const auto iter = std::find(m_data.begin(), m_data.end(), value);
-	if (iter != m_data.end())
-	{
-		return std::distance(m_data.begin(), iter);
-	}
-	else
-	{
-		return -1;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::clear()
-{
-	m_data.clear();
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::erase(int pos)
-{
-	m_data.erase(m_data.begin() + pos);
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::insert(int pos, const std::string& value)
-{
-	m_data.insert(m_data.begin() + pos, value);
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::push_back(const std::string& value)
-{
-	m_data.push_back(value);
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::resize(int count)
-{
-	m_data.resize(count);
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::set(int pos, const std::string& value)
-{
-	m_data[pos] = value;
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::swap(int i, int j)
-{
-	if (i < 0 || i >= size() || j < 0 || j >= size())
-	{
-		return;
-	}
-
-	std::swap(m_data[i], m_data[j]);
-	m_modified = true;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::load(XfceRc* rc, bool is_default)
-{
-	if (!xfce_rc_has_entry(rc, m_property + 1))
-	{
-		return;
-	}
-
-	gchar** data = xfce_rc_read_list_entry(rc, m_property + 1, ",");
-	if (!data)
-	{
-		return;
-	}
-
-	std::vector<std::string> strings;
-	for (int i = 0; data[i]; ++i)
-	{
-		strings.push_back(data[i]);
-	}
-	set(strings, !is_default);
-
-	g_strfreev(data);
-
-	if (is_default)
-	{
-		m_default = m_data;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-bool StringList::load(const gchar* property, const GValue* value, bool& reload_menu)
-{
-	if (g_strcmp0(m_property, property) != 0)
-	{
-		return false;
-	}
-
-	// Ignore own changes to prevent extra menu reload
-	if (m_saved)
-	{
-		m_saved = false;
-		reload_menu = false;
-		return true;
-	}
-
-	// Handle resetting to default
-	if (G_VALUE_TYPE(value) == G_TYPE_INVALID)
-	{
-		m_modified = false;
-		m_order_unchanged = false;
-		m_data = m_default;
-		reload_menu = true;
-		return true;
-	}
-
-	// Convert GValue to string list
-	std::vector<std::string> strings;
-	if (G_VALUE_HOLDS(value, G_TYPE_PTR_ARRAY))
-	{
-		const GPtrArray* values = static_cast<const GPtrArray*>(g_value_get_boxed(value));
-		for (guint i = 0; i < values->len; ++i)
-		{
-			const GValue* string = static_cast<const GValue*>(g_ptr_array_index(values, i));
-			if (G_VALUE_HOLDS_STRING(string))
-			{
-				strings.push_back(g_value_get_string(string));
-			}
-		}
-	}
-	else if (G_VALUE_HOLDS(value, G_TYPE_STRV))
-	{
-		const gchar** values = static_cast<const gchar**>(g_value_get_boxed(value));
-		for (int i = 0; values[i]; ++i)
-		{
-			strings.push_back(values[i]);
-		}
-	}
-	else if (G_VALUE_HOLDS_STRING(value))
-	{
-		strings.push_back(g_value_get_string(value));
-	}
-
-	// Load string list
-	set(strings, false);
-	reload_menu = true;
-
-	return true;
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::save()
-{
-	if (!m_modified || !m_settings->channel)
-	{
-		return;
-	}
-
-	m_settings->begin_property_update();
-
-	const int size = m_data.size();
-	GPtrArray* array = g_ptr_array_sized_new(size);
-
-	for (int i = 0; i < size; ++i)
-	{
-		GValue* value = g_new0(GValue, 1);
-		g_value_init(value, G_TYPE_STRING);
-		g_value_set_static_string(value, m_data[i].c_str());
-		g_ptr_array_add(array, value);
-	}
-
-	xfconf_channel_set_arrayv(m_settings->channel, m_property, array);
-	xfconf_array_free(array);
-
-	m_saved = true;
-	m_modified = false;
-
-	m_settings->end_property_update();
-}
-
-//-----------------------------------------------------------------------------
-
-void StringList::set(std::vector<std::string>& data, bool store)
-{
-	m_data.clear();
-
-	for (auto& desktop_id : data)
-	{
-		if (desktop_id == "exo-web-browser.desktop")
-		{
-			desktop_id = "xfce4-web-browser.desktop";
-		}
-		else if (desktop_id == "exo-mail-reader.desktop")
-		{
-			desktop_id = "xfce4-mail-reader.desktop";
-		}
-		else if (desktop_id == "exo-file-manager.desktop")
-		{
-			desktop_id = "xfce4-file-manager.desktop";
-		}
-		else if (desktop_id == "exo-terminal-emulator.desktop")
-		{
-			desktop_id = "xfce4-terminal-emulator.desktop";
-		}
-		if (std::find(m_data.begin(), m_data.end(), desktop_id) == m_data.end())
-		{
-			m_data.push_back(std::move(desktop_id));
-		}
-	}
-
-	m_modified = store;
-	m_order_unchanged = false;
-}
-
-//-----------------------------------------------------------------------------
-
-SearchActionList::SearchActionList(WhiskerMenu::Settings* settings, std::initializer_list<SearchAction*> data) :
-	m_settings(settings),
-	m_data(data),
-	m_modified(false)
-{
-	clone(m_data, m_default);
-}
-
-//-----------------------------------------------------------------------------
-
-SearchActionList::~SearchActionList()
-{
-	for (auto action : m_default)
-	{
-		delete action;
-	}
-
-	for (auto action : m_data)
-	{
-		delete action;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-void SearchActionList::erase(SearchAction* value)
-{
-	m_data.erase(std::find(m_data.begin(), m_data.end(), value));
-	m_modified = true;
-}
-
-//-----------------------------------------------------------------------------
-
-void SearchActionList::push_back(SearchAction* value)
-{
-	m_data.push_back(value);
-	m_modified = true;
-}
-
-//-----------------------------------------------------------------------------
-
-void SearchActionList::load(XfceRc* rc, bool is_default)
-{
-	const int size = xfce_rc_read_int_entry(rc, "search-actions", -1);
-	if (size < 0)
-	{
-		return;
-	}
-
-	for (int i = 0; i < size; ++i)
-	{
-		gchar* key = g_strdup_printf("action%i", i);
-		if (!xfce_rc_has_group(rc, key))
-		{
-			g_free(key);
-			continue;
-		}
-		xfce_rc_set_group(rc, key);
-		g_free(key);
-
-		SearchAction* action = new SearchAction(m_settings,
-				xfce_rc_read_entry(rc, "name", ""),
-				xfce_rc_read_entry(rc, "pattern", ""),
-				xfce_rc_read_entry(rc, "command", ""),
-				xfce_rc_read_bool_entry(rc, "regex", false));
-
-		bool found = false;
-		for (auto current : m_data)
-		{
-			if (*current == *action)
-			{
-				found = true;
-				break;
-			}
-		}
-
-		if (!found)
-		{
-			m_data.push_back(action);
-			m_modified = true;
-		}
-		else
-		{
-			delete action;
-		}
-	}
-
-	if (is_default)
-	{
-		clone(m_data, m_default);
-		m_modified = false;
-	}
-}
-
-//-----------------------------------------------------------------------------
-
-void SearchActionList::load()
-{
-	const int size = xfconf_channel_get_int(m_settings->channel, "/search-actions", -1);
-	if (size < 0)
-	{
-		return;
-	}
-
-	for (auto action : m_data)
-	{
-		delete action;
-	}
-	m_data.clear();
-
-	gchar* property = nullptr;
-	gchar* name = nullptr;
-	gchar* pattern = nullptr;
-	gchar* command = nullptr;
-	bool regex;
-
-	for (int i = 0; i < size; ++i)
-	{
-		property = g_strdup_printf("/search-actions/action-%d/name", i);
-		name = xfconf_channel_get_string(m_settings->channel, property, nullptr);
-		g_free(property);
-
-		property = g_strdup_printf("/search-actions/action-%d/pattern", i);
-		pattern = xfconf_channel_get_string(m_settings->channel, property, nullptr);
-		g_free(property);
-
-		property = g_strdup_printf("/search-actions/action-%d/command", i);
-		command = xfconf_channel_get_string(m_settings->channel, property, nullptr);
-		g_free(property);
-
-		property = g_strdup_printf("/search-actions/action-%d/regex", i);
-		regex = xfconf_channel_get_bool(m_settings->channel, property, false);
-		g_free(property);
-
-		m_data.push_back(new SearchAction(m_settings, name, pattern, command, regex));
-
-		g_free(name);
-		g_free(pattern);
-		g_free(command);
-	}
-
-	m_modified = false;
-}
-
-//-----------------------------------------------------------------------------
-
-bool SearchActionList::load(const gchar* property, const GValue* value)
-{
-	if (g_strcmp0("/search-actions", property) == 0)
-	{
-		if (G_VALUE_TYPE(value) != G_TYPE_INVALID)
-		{
-			load();
-		}
-		else
-		{
-			clone(m_default, m_data);
-		}
-		return true;
-	}
-
-	int index = 0;
-	char field[16];
-	if (std::sscanf(property, "/search-actions/action-%d/%14s", &index, field) != 2)
-	{
-		return false;
-	}
-
-	if (index >= size())
-	{
-		return true;
-	}
-	SearchAction* action = m_data[index];
-
-	if ((g_strcmp0(field, "name") == 0) && G_VALUE_HOLDS_STRING(value))
-	{
-		action->set_name(g_value_get_string(value));
-	}
-	else if ((g_strcmp0(field, "pattern") == 0) && G_VALUE_HOLDS_STRING(value))
-	{
-		action->set_pattern(g_value_get_string(value));
-	}
-	else if ((g_strcmp0(field, "command") == 0) && G_VALUE_HOLDS_STRING(value))
-	{
-		action->set_command(g_value_get_string(value));
-	}
-	else if ((g_strcmp0(field, "regex") == 0) && G_VALUE_HOLDS_BOOLEAN(value))
-	{
-		action->set_is_regex(g_value_get_boolean (value));
-	}
-
-	return true;
-}
-
-//-----------------------------------------------------------------------------
-
-void SearchActionList::save()
-{
-	if (!m_modified || !m_settings->channel)
-	{
-		return;
-	}
-
-	m_settings->begin_property_update();
-
-	xfconf_channel_reset_property(m_settings->channel, "/search-actions", true);
-
-	const int size = m_data.size();
-	xfconf_channel_set_int(m_settings->channel, "/search-actions", size);
-
-	gchar* property = nullptr;
-	const SearchAction* action = nullptr;
-
-	for (int i = 0; i < size; ++i)
-	{
-		action = m_data[i];
-
-		property = g_strdup_printf("/search-actions/action-%d/name", i);
-		xfconf_channel_set_string(m_settings->channel, property, action->get_name());
-		g_free(property);
-
-		property = g_strdup_printf("/search-actions/action-%d/pattern", i);
-		xfconf_channel_set_string(m_settings->channel, property, action->get_pattern());
-		g_free(property);
-
-		property = g_strdup_printf("/search-actions/action-%d/command", i);
-		xfconf_channel_set_string(m_settings->channel, property, action->get_command());
-		g_free(property);
-
-		property = g_strdup_printf("/search-actions/action-%d/regex", i);
-		xfconf_channel_set_bool(m_settings->channel, property, action->get_is_regex());
-		g_free(property);
-	}
-
-	m_modified = false;
-
-	m_settings->end_property_update();
-}
-
-//-----------------------------------------------------------------------------
-
-void SearchActionList::clone(const std::vector<SearchAction*>& in, std::vector<SearchAction*>& out) const
-{
-	// Remove previous actions
-	for (auto action : out)
-	{
-		delete action;
-	}
-	out.clear();
-
-	// Copy actions
-	out.reserve(in.size());
-	for (auto action : in)
-	{
-		out.push_back(new SearchAction(m_settings,
-				action->get_name(),
-				action->get_pattern(),
-				action->get_command(),
-				action->get_is_regex()));
-	}
-}
-
-//-----------------------------------------------------------------------------
-
 const std::vector<std::string>& Settings::get_aliases(const char* desktop_id) const
 {
 	static const std::vector<std::string> empty;
@@ -1305,165 +620,3 @@ void Settings::save_aliases(XfconfChannel* ch)
 	end_property_update();
 }
 
-//-----------------------------------------------------------------------------
-
-/* migrate_schema:
- * @is_fresh_install: true when no Xfconf properties were present at load.
- *
- * Walks the channel forward through every known schema version, applying
- * additive migrations. Versions are cumulative: each block runs once per
- * upgrade. See .specify/specs/003-properties-refactor/contracts/xfconf-keys.md
- * for the v2 contract.
- */
-void Settings::migrate_schema(bool is_fresh_install)
-{
-	if (!channel)
-		return;
-
-	const int current_schema = 3;
-	if (schema_version >= current_schema)
-		return;
-
-	begin_property_update();
-
-	if (schema_version < 1)
-	{
-		// Map legacy menu-opacity → categories-opacity if present and categories-opacity missing
-		if (xfconf_channel_has_property(channel, "/menu-opacity")
-				&& !xfconf_channel_has_property(channel, "/categories-opacity"))
-		{
-			const int legacy_opacity = xfconf_channel_get_int(channel, "/menu-opacity", 100);
-			xfconf_channel_set_int(channel, "/categories-opacity", legacy_opacity);
-			categories_opacity = legacy_opacity;
-		}
-
-		// Write defaults for V1 properties not yet in the channel
-		struct { const char* prop; int val; } int_props[] = {
-			{ "/corner-radius",       0   },
-			{ "/panel-gap",           0   },
-			{ "/categories-opacity",  100 },
-			{ "/apps-opacity",        100 },
-			{ "/grid-columns",        4   },
-			{ "/grid-rows",           3   },
-		};
-		for (auto& p : int_props)
-		{
-			if (!xfconf_channel_has_property(channel, p.prop))
-				xfconf_channel_set_int(channel, p.prop, p.val);
-		}
-
-		// NOTE: defaults for V1 Boolean properties not yet in the channel;
-		// /unified-bar default false matches the C++ Settings ctor and lets
-		// downstream consumers read the key explicitly after first migration.
-		struct { const char* prop; gboolean val; } bool_props[] = {
-			{ "/unified-bar", FALSE },
-		};
-		for (auto& p : bool_props)
-		{
-			if (!xfconf_channel_has_property(channel, p.prop))
-				xfconf_channel_set_bool(channel, p.prop, p.val);
-		}
-
-		struct { const char* prop; const char* val; } str_props[] = {
-			{ "/sidebar-position",     "left"      },
-			{ "/search-bar-position",  "top"       },
-			{ "/profile-position",     "top"       },
-			{ "/commands-position",    "top-right" },
-			{ "/grid-density",         "medium"    },
-			{ "/layout-mode",          "docked"    },
-		};
-		for (auto& p : str_props)
-		{
-			if (!xfconf_channel_has_property(channel, p.prop))
-				xfconf_channel_set_string(channel, p.prop, p.val);
-		}
-
-		if (is_fresh_install)
-			apply_preset(BUILTIN_PRESETS[PRESET_MODERN], *this);
-		else
-			current_preset_id = "classic";
-
-		schema_version = 1;
-	}
-
-	if (schema_version < 2)
-	{
-		// Seed /full-screen-opacity (new key). Default 100; honour active preset if it pins one.
-		// NOTE: at this point preset.cpp's file-seeded values are not yet read — the active
-		// preset's compiled default is used as the fallback. Users who want a custom value
-		// can edit it via the Properties dialog after upgrade.
-		if (!xfconf_channel_has_property(channel, "/full-screen-opacity"))
-			xfconf_channel_set_int(channel, "/full-screen-opacity", 100);
-
-		// Deprecate /position-categories-horizontal: subsumed by /sidebar-position ∈ {top, bottom}.
-		// If the user had it on AND sidebar-position is left|right (or unset), promote sidebar-position
-		// to "top". Then reset the dead key to false so future reads are inert.
-		if (xfconf_channel_has_property(channel, "/position-categories-horizontal")
-				&& xfconf_channel_get_bool(channel, "/position-categories-horizontal", false))
-		{
-			gchar* current_sidebar = xfconf_channel_get_string(channel, "/sidebar-position", nullptr);
-			const bool sidebar_is_vertical = !current_sidebar
-					|| g_strcmp0(current_sidebar, "left") == 0
-					|| g_strcmp0(current_sidebar, "right") == 0;
-			if (sidebar_is_vertical)
-			{
-				xfconf_channel_set_string(channel, "/sidebar-position", "top");
-				sidebar_position = "top";
-			}
-			g_free(current_sidebar);
-			xfconf_channel_set_bool(channel, "/position-categories-horizontal", false);
-			position_categories_horizontal = false;
-		}
-
-		// Deprecate profile-shape = Hidden (enum value 2): subsumed by /profile-position = "hidden".
-		// HACK: we keep the ProfileHidden enum value compiled in to avoid breaking any third-party
-		// preset file that still references the integer 2, but the UI no longer exposes it.
-		if (xfconf_channel_has_property(channel, "/profile-shape")
-				&& xfconf_channel_get_int(channel, "/profile-shape", ProfileRound) == ProfileHidden)
-		{
-			xfconf_channel_set_string(channel, "/profile-position", "hidden");
-			profile_position = "hidden";
-			xfconf_channel_set_int(channel, "/profile-shape", ProfileRound);
-			profile_shape = ProfileRound;
-		}
-
-		schema_version = 2;
-	}
-
-	if (schema_version < 3)
-	{
-		// Milestone 005 — Places mode keys. Seed defaults on first upgrade so
-		// existing installs see consistent values without re-applying a preset.
-		struct { const char* prop; gboolean val; } bool_props[] = {
-			{ "/places/enabled",            FALSE },
-			{ "/places/history-enabled",    TRUE  },
-			{ "/places/favourites-enabled", TRUE  },
-			{ "/places/remember-last-mode", FALSE },
-			{ "/places/show-metadata",      FALSE },
-		};
-		for (auto& p : bool_props)
-		{
-			if (!xfconf_channel_has_property(channel, p.prop))
-				xfconf_channel_set_bool(channel, p.prop, p.val);
-		}
-
-		struct { const char* prop; const char* val; } str_props[] = {
-			{ "/places/favourite-sync", "meowmenu" },
-			{ "/places/last-mode",      "apps"     },
-		};
-		for (auto& p : str_props)
-		{
-			if (!xfconf_channel_has_property(channel, p.prop))
-				xfconf_channel_set_string(channel, p.prop, p.val);
-		}
-
-		if (!xfconf_channel_has_property(channel, "/places/max-items"))
-			xfconf_channel_set_int(channel, "/places/max-items", 20);
-
-		schema_version = 3;
-	}
-
-	end_property_update();
-}
-
-//-----------------------------------------------------------------------------
